Refactor dashboard into components
Splits the dashboard page into smaller, more maintainable components to improve code organization and readability.
